Added Matrix library
Added the npm "ml-matrix" library to assist with verifying answers to some of the linear algebra questions.

This required the use of Browserify, which means that other npm packages can be added later, if desired.

The use of Browserify also required us to switch from the JavaScript import/export system to the Node require/module.exports system.  These systems are mostly identical though, so it didn't change much.

Added a Linear Algebra subject as a proof of concept and test of following the instructions I had written for creating a new subject.