Added Matrix Answer Type
Added an answer type for matrices for the new Linear Algebra subject.
This required changing the structure of the AnswerType "enum" from an object to a class, but I was able to preserve backwards compatibility.
The answer type can be used as a fixed matrix size or a variable size (depending on the matrices created in the question).
The answer type handling code is now more complex, but not overly so, because we can (currently) differentiate between matrix and non-matrix types by checking if the type is an array.