Switched Matrix Library
Switched the matrix library we use from ml-matrix to math.js.  I did this because I found a bug in ml-matrix's rref function (it didn't handle negative values correctly).  This caused me to lose confidence in the library as a whole, so I switched to math.js.
Additionally, I was probably going to use math.js's complex number support at some point, so this was a good time to get it set up anyway.

Added some stuff to Linear Algebra using the matrix library.  I can confirm that it works well for our purposes.