Improved Linear Algebra
Added a new answer type, Vectors, that compares sets of vectors (order is ignored).  This was made easier by redoing the way the answer type object stores data for matrix answer types.  Now the relevant dimensional data is not stored in the answer type, but is given to the question later (when constant functions are resolved).  This makes much of the special case logic we had to do for matrices earlier much nicer (i.e. we can test directly for answerType === matrix, instead of testing if it's an array).
I also made the first custom validator in this project (I used it to compare bases of vector spaces) and discovered some things that needed to be changed about how custom validators are implemented.  I documented these features in the ReadMe.

Corrected a bug in the numeric parser, so that it can now accept numbers as input not just strings.

Halfway point on the Linear Algebra subject (Linear Systems and Vector Spaces are finished, only Eigenvalues and Inner Products remain).