Added Complex Variables
Added the complex variables subject.  In the process, I made several changes and improvements to parts of the code.  As always, backwards compatibility should be maintained, but I have really tested it.

I redid how the problem importance level is selected so that it takes into account the number of problems in each level.
I (hopefully) fixed the answer input label text, so a colon should appear before the default label (when used), but custom labels will not have it automatically added. (Note, this may require going through the Lin Alg and DiffEq subjects to correct any differences in this policy).

I repurposed the matrix_dims field of the questions so that it has the more general purpose of providing extra information about the answer type, whatever it may be (This is now used in 5 different answer types [if desired]).

Updated the correct answer display for MultipleChoice and SelectAll questions so that it now shows the display value rather than the key since the key holds no meaning for the user.

Added an expression and series answer types.  This required adding an expression parser to parse latex.  The parser works well, but the expression validator doesn't.  The expression validator makes use of math.js's symbolic expression checker which doesn't work very well for our purposes.  (We had to add a check for equal strings because the checker didn't recognize *equal* strings as being the same expression).

Also, complex variables is the first subject to have theorem flashcard type things.  Basically, the question gives some information about the theorem, then the user selects the rest from a multiple choice thing.